requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
psycopg-pool
python-dotenv
pyjwt
passlib[bcrypt]
argon2-cffi
langchain-google-genai
langgraph
langchain-chroma
langchain-mcp-adapters
fastmcp

PyPortfolioOpt
tavily-python
pandas
numpy
yfinance
requests
pydantic
pandas-ta
gTTS

